* support for GTK on FreeBSD:  -> treat <GType> as a <machine-word> (because it uses more than 29 bits)  -> define constants for G-TYPE-* in gtk.dylan, remove from gtk-c-ffi  -> remove several unused bindings from gtk-c-ffi,       which are undefined on FreeBSD  -> registry entries for gtk, duim, gtk-c-ffi for FreeBSD
* export <C-buffer-offset>, <LPSOCKADDR> and <timeval> from network library
  (these are also exported on Windows, no reason why they should not be
   exported on UNIX, too (needed in network-interfaces))

+++ sources/lib/run-time/linux-types.h
@@ -10,7 +10,7 @@ typedef long                *LPLONG;
 typedef unsigned char        BYTE;
 typedef int                  BOOL;
 typedef unsigned long        DWORD;
-typedef int64_t              _int64;
+typedef long long            _int64;
 
 #define WINAPI
